Sketcher: getSubObject support for internal geometries
When editing, the selected sub-element name will be appended with '_' to
tell SketchObject::getSubObject() that the element is referring to
internal geometry. This makes it possible to retrieve the shape of
internal geometries, such as constructions or external geometries.

This patch also explicitly set the editing document, object and subname
when doing selection in 3D view. Because of the support of in-place
external editing, the editing may originated from various child/link
inside some container. This patch makes sure that during editing,
selection in 3D view will cause TreeView to select the correct item.

Note, there is no support for enhanced selection from other source
during editing yet, e.g. from various commands or task panels.