Sketcher: Geometry Extension for Internal Alignment Geometry
============================================================

This commit extends SketchGeometryExtension to store within a Geometry if it is an internal aligment geometry
and which type.

It extends the restore mechanism so that a construction point of the sketcher is restored as an BSpline-Knot, thereby
decoupling the construction status from this functionality.

BSpline handling code will be adapted in the next commit to separate the implementation of the extension from the use.