Sketcher: Use geometry extension Internal Types on geometry creation
====================================================================

What?

SketchGeometryExtension provides storage to know whether a geometry
is an internal geometry (geometry used via Internal Alignment mechanism
to constraint another geometry, such as bspline poles or knots).

This commit extends ExposeInternalGeometry to use this facility. At this
moment the most important is to store the BSplineKnot status for solving
reasons, others are planned to be used for graphical representation
enhancement.

This commit extends addGeometry Python version to take an optional third
additional parameter to directly create a geometry with the right InternalType
information. This is used in CommandCreateGeo for the creation of B-Spline poles.