Sketcher: After restore migration mechanism
===========================================

The migration mechanism into extensions is currently based on the trigger that no
SketchGeometryExtension is present for a given Geometry.

It has two phases:

Phase 1: uses SketchObject::handleSpecialProperty to tackle restore of previously used storage schemes in
	 Part namespace. This is the case of LinkStage3. It could be the case too if Construction is moved
	 to Sketcher.

Phase 2: When new extension information requires access to other properties (for example Constraints), which
	 either have not been restored, or simply migration would require code that do not belong functionaly
	 to specially restoring a property, a flag afterRestoreMigration is set, which triggers migration from
	 SketchObject::onDocumentRestored()

This commit is Phase 2.