Sketcher: TextGeometryExtension
===============================

Extension intended to bring text along a path defined by a geometry in the Sketcher.

It has a toShape() method that produces:
- An empty TopoDS_Shape is member shapeEnabled is false.
- Either a TopoDS_Wire or a TopoDS_Compound embedded in a TopoDS_Shape if member shapeEnabled is true.

TopoDS_Wire is for the special case that the shape can be produced by a single wire.

TopoDS_Compound MUST be a list of TopoDS_Wires and covers the general case where several wires are necessary
to produce the text shape.