Path: Relocate `valueChanged` signal translation to QSB class
Moved the translation method from PathFeatureExtensionsGui module to proper location within QuantitySpinBox class in PathGui module.
This change adds a missing translation method for passing the `editingFinished` signal to the parent task panel when the user is finished editing a QuantitySpinBox object with an active expression in the Gui.
If no expression is active, this new method is dormant.

Some cleanup of initial changes related to development of this fix are included.