Console/ILogger: Refactor and extension
=======================================

Refactor:
- Substitute the use of variadic templates with parameter packs.
- Use external library "fmt" as header only to handle printf like formating.

Extension:
- Include a notifier field, so that the originator of the message can be provided. E.g. Document#DocumentObject
- Include a new type of message called CriticalMessage, this message is intended to have
special behaviour in the future. Namely, it will be used to notify forward compatilibity issues.
It will be used to substitute the current signal/slot mechanism.
- Include two new types of messages for user notifications (Notification and TranslatedNotification). This messages
will be use to convey UI notifications intended for the user (such as non-intrusive message about the usage of a tool). There
are two versions to mark whether the string provided as a message is already translated or not. When using the console system for
notifications, these notifications may originate from the App or the Gui. In the former, it is generally the case that the strings
of messages are not (yet) translated (but they can be marked with QT_TRANSLATE_NOOP). In the latter, often the messages to be provided
are already translated.

Library requirements:
- Libfmt must be installed (will need to be added to the libpack).
sudo apt-get install libfmt-dev

Python support for CriticalMessage, Notification and TranslatedNofification, including shortcuts:

Crt = FreeCAD.Console.PrintCriticalMessage
Ntf = FreeCAD.Console.PrintNotification
Tnf = FreeCAD.Console.PrintTranslatedNotification