App: DocumentObject - New full label member function
====================================================

Before this commit there is a function getFullName(), which
produces a string in the form document#objectname. For "document"
the real name of the document is taken, whereas for objectname it
is the initial name (Sketch, Sketch001), not the label. If the
real name of the document is in foreign languages, such as JP, it
is not correctly coded (document name appears as underscores).

The new function fo this commit, getFullLabel(), provides a string
documentlabel#documentobjectlabel, which is properly coded for
foreign languages and has the advantage to use labels, that are
probably easier to identify for end users.